requirements-apple-m1.txt
tqdm==4.55.1
pandas==1.2.4
pycocotools==2.0.2
wandb
Pillow>=7.1.2

torch==1.9.0
torchvision==0.10.0